Extract `inlineAllContent` and place it into `data` package.
