+++ packages/data/src/index.ts
@@ -25,3 +25,5 @@ export {
 } from "./inflate-deflate.js";
 
 export * from "./common.js";
+
+export { inlineAllContent } from "./inline-all-content.js";
